Rename uvu-timeout.ts to rig-test.ts
