+++ src/test/cache-common.ts
@@ -6,7 +6,7 @@
 
 import * as assert from 'uvu/assert';
 import * as pathlib from 'path';
-import {rigTest} from './util/uvu-timeout.js';
+import {rigTest} from './util/rig-test.js';
 import {sep} from 'path';
 import {checkScriptOutput} from './util/check-script-output.js';
