+++ src/test/clean.test.ts
@@ -6,7 +6,7 @@
 
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
-import {rigTest} from './util/uvu-timeout.js';
+import {rigTest} from './util/rig-test.js';
 import * as pathlib from 'path';
 import {checkScriptOutput} from './util/check-script-output.js';
